requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3

paho-mqtt~=1.6.1
opencv-python~=4.7.0.72
numpy~=1.24.3
Pillow~=9.5.0